final lab, integrate weather api

+++ consumer/src/main/java/miu/edu/bdt/consumer/dto/Location.java
@@ -13,8 +13,8 @@ public class Location {
     private String name;
     private String region;
     private String country;
-    private double lat;
-    private double lon;
+    private float lat;
+    private float lon;
     private String tz_id;
     private long localtime_epoch;
     private String localtime;
